requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.9
django-extensions==1.9.9
djangorestframework==3.3.0
idna==2.6
lxml==4.1.1
requests==2.18.4
six==1.11.0
typing==3.6.2
urllib3==1.22
